requirements.txt
PyGithub==1.25.2
SQLAlchemy==0.8.1
nose==1.3.0
requests==1.2.0
unicodecsv==0.9.4
ghp-import==0.2.2
